requirements.txt
numpy==1.23.4
opencv_contrib_python==4.6.0.66
Pillow==9.3.0
PyAutoGUI==0.9.53
pynput==1.7.6
